fix(checkoutService): patch shippingAddress in modelCopy
Previously the shippingAddress was set to the
billingAddress directly in the checkoutModel
which breaks the app code since then both
models point to the same reference. The app
could have worked around that by manually
switching references back and forth but
it seems to be cleaner to just let the
checkoutService do it's manupulations on
the modelCopy instead of manipulating the
checkoutModel directly.